'finished' analysis and repl
