+++ analyze_commits.py
@@ -1,6 +1,7 @@
 """Execute analysis for commits."""
 import os
 
+
 from emoji_library import get_emojis_count
 from analyze_sentiment import get_avg_sentiment
 import parse_commits
@@ -10,33 +11,37 @@ from dulwich import porcelain
 
 
 def analyze_commits(out):
-    """Execute analysis for commits."""
+    """Execute analysis for commits."""   
+    return_list = []
+    return_list = get_list_of_commits(out)
+    write_string = ""
+    print("Number of commits: " + str(len(return_list)))
+    write_string += "Number of commits: " + str(len(return_list))
+
+    gg_gensim.gensim_analysis(return_list)
+
+    sentiment = get_avg_sentiment(return_list)
+    for key, value in sentiment.items():
+        print(key, value)
+        write_string += str(key) + ", " + str(value) + "\n"
+
+    emojis_count = get_emojis_count(return_list)
+    for key, value in emojis_count.items():
+        print(key, value)
+        write_string += str(key) + ", " + str(value) + "\n"
+    print(write_string)
+    # embed_stats_into_html(write_string)
+    
+    
+def get_list_of_commits(out):
+    """Write a list of commit messages."""
     repo_list = next(os.walk("./" + str(out)))[1]
+    return_list = []
     for repo in repo_list:
-        return_list = []
-        with porcelain.open_repo_closing("./" + str(out) + "/" + str(repo)) as repo:
+        with porcelain.open_repo_closing("./" + str(out) + "/"+ str(repo)) as repo:
             walker = repo.get_walker(reverse=True)
         for entry in walker:
             item = str(entry.commit.message.decode())
             item = item.replace("\n", "")
             return_list.append(item)
-
-    list_of_commits = parse_commits.get_list_of_commits(out)
-
-    write_string = ""
-    print("Number of commits: " + len(list_of_commits))
-    write_string += "Number of commits: " + len(list_of_commits)
-
-    gg_gensim.gensim_analysis(list_of_commits)
-
-    sentiment = get_avg_sentiment(list_of_commits)
-    for key, value in sentiment.iteritems():
-        print(key, value)
-        write_string += key + ", " + value + "\n"
-
-    emojis_count = get_emojis_count(list_of_commits)
-    for key, value in emojis_count.iteritems():
-        print(key, value)
-        write_string += key + ", " + value + "\n"
-
-    embed_stats_into_html(write_string)
+    return return_list
